chore(ui5-calendar): remove calendar header component
