+++ packages/main/src/Calendar.hbs
@@ -51,25 +51,9 @@
 			@ui5-navigate="{{onNavigate}}"
 		></ui5-yearpicker>
 	</div>
-
-	{{! Positioned above the content due to flex-direction:column-reverse, but physically here for tab order to work}}
-	<ui5-calendar-header
-			id="{{_id}}-head"
-			.primaryCalendarType="{{_primaryCalendarType}}"
-			.secondaryCalendarType="{{_secondaryCalendarType}}"
-			.buttonTextForSecondaryCalendarType="{{secondaryCalendarTypeButtonText}}"
-			timestamp="{{_timestamp}}"
-			.isPrevButtonDisabled="{{_previousButtonDisabled}}"
-			.isNextButtonDisabled="{{_nextButtonDisabled}}"
-			.isMonthButtonHidden="{{_isHeaderMonthButtonHidden}}"
-			._monthButtonText="{{_headerMonthButtonText}}"
-			._yearButtonText="{{_headerYearButtonText}}"
-			._yearButtonTextSecType="{{_headerYearButtonTextSecType}}"
-			@ui5-previous-press="{{onHeaderPreviousPress}}"
-			@ui5-next-press="{{onHeaderNextPress}}"
-			@ui5-show-month-view="{{onHeaderShowMonthPress}}"
-			@ui5-show-year-view="{{onHeaderShowYearPress}}"
-	></ui5-calendar-header>
+	<div class="ui5-calheader">
+		{{>include "./CalendarHeader.hbs"}}
+	</div>
 </div>
 
 <div
